==> 02_Song_Recommender_Create_Song_Database.ipynb <==
# Creating a song database using the Spotify API 

These functions can be used to build a new spotify song database that can later be used for the Spotify recommender.<br>
We have created a song database and saved it to a .csv file which we will use later for the recommender, but these functions can be used to build a database from scratch. There is also some dictionaries of playlists that can be used for it. 

import config

import pandas as pd
import spotipy
import json
import pprint
from spotipy.oauth2 import SpotifyClientCredentials

sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id= config.client_id,
                                                           client_secret= config.client_secret))

## Get tracks from a playlist

This function returns a list of tracks from a Spotify playlist, it will be used later inside of another function.

def get_playlist_tracks(username, playlist_id):
    results = sp.user_playlist_tracks(username,playlist_id,market="GB")
    tracks = results['items']
    while results['next']:
        results = sp.next(results)
        tracks.extend(results['items'])
    return tracks

## Create a dataframe with audiofeatures

This function creates a dataframe from a given playlist - the arguments are playlist_ID and user_name of the creator.

def create_df_with_audio_features (playlist_id,username): 
    
    '''as arguments pass a spotify playlist id and the corresponding user name. 
    the function returns a dataframe with audiofeatures of all songs from given playlist.'''
    
    tracks = get_playlist_tracks(username, playlist_id) 
    
    track_id_list = []

    for p in range (len(tracks)):
        track_id_list.append(tracks[p]['track']["id"])
    
    audio_features = sp.audio_features(track_id_list) # this function takes maximum of 100 id's !!!
    
    songs_dataframe = pd.DataFrame(audio_features)
    
    return songs_dataframe

## Add songs from playlists to a dataframe

This function uses a dictionary of playlists to add new songs to an existing dataframe (don't forget to remove duplicates in the end, because songs might be repeated on different playlists!)

def create_data_frame_from_multiple_dataframes(playlist_id_and_username_dict, songs_audio_features_df):
    
    '''as arguments pass a dictionary where keys are playlist ids and values are usernames 
    and an existing data frame of song audio features'''
    
    for key, value in playlist_id_and_username_dict.items():
        x = create_df_with_audio_features(key,value)
        songs_audio_features_df = pd.concat(objs=[songs_audio_features_df, x], axis=0,ignore_index=True)
        
    songs_audio_features_df.to_csv("Data/songs_audio_feature_dataframe1.csv", index=False)
                               
    return songs_audio_features_df

## Example dictionaries of playlists

new_dict = {'0j8LmLIBQsFENpw7xATmRQ':'Tony Wiese','0zX2XXdBjwxWkEJh8GnTbK':'Marcus Bissetti'}
playlist_dict_4 ={'37i9dQZF1DX6RA5ZrA5a23':'Spotify','37i9dQZF1DWXWbLEOaHnU3':'Spotify'}
playlist_dict_5 = {'37i9dQZF1DXbS8bPVXXR2B':'Spotify','37i9dQZF1DX4arVIN5Cg4U':'Spotify','37i9dQZF1DX8cbjeZ4UYIG':'Spotify'}
velvey_dict_1 = {'63OX6biHoyrdDLIa6wSzYC':'velvey','2WTgjtdNpP5eINfT0nVu4R':'velvey','2nr5kV7ruSHveM5OZWdHIb':'velvey','62ph81KhF22tCbbH4OA4gM':'velvey','36GzxsljeKxdsj3WVUBSUl':'velvey'}
velvey_dict_2 = {'6msAESikByV5t2PPQBBJse':'velvey','24qEcbVYRIQ3GYhFl7Tu3g':'velvey','46RbwjSzFjBqX1zXQwq2V2':'velvey','1hYS1PLac5pNr1AGitRyG9':'velvey','0JOKEhfKGOxfaQZt9QUv7i':'velvey'}
velvey_dict_4 = {'3WkGqIO7lWsbpj0vIQzvIQ':'velvey','75Qx7bUKAkxb9XkD3rSKEt':'velvey','4KXQ8JAP0OkTASg41S9B5n':'velvey','1IJmQ1ddrzWVYIHPwGMvGg':'velvey'}
velvey_dict_3 = {'5kexhZdmEHsybU97DjAC0F':'velvey','0RxyjQBJsfx0QSS887gfGd':'velvey','7cNF6tUhKlXWuw0JVKLZwA':'velvey','7CGbc91x2JU3mddaEJlTEk':'velvey'}
nelson_playlist = {'4ldLuRieh8jnCKOjSoqUQH':'Nelson Lage'}
magdas_playlists = {'71nFwFmZFvVYmils0i0KS6':'Magdalena Blasky','37i9dQZF1DX3jWba5xiDhV':'Spotify'}
jurek_playlists = {'3rK9vLarAA4KNZr0DuCcEJ':'jerry_woo','74baqx6L1qmWf6iU5vVMAi':'jerry_woo','4k4FbGz5QUb3Wkqd5B4hBm':'jerry_woo','6Eh8CSnguxXALsM4M6YsQS':'jerry_woo'}
filips_playlists_1 = {'5aP6mwRVDOVNGYrrZQCxE8':'Filip Piotr Makara'}
filips_playlists_2 = {'40cdOFiijrki5AcTQEjPnR':'Filip Piotr Makara'}
filips_playlists_3 = {'5OW5iLQJXKA7udpjlxxjCG':'Filip Piotr Makara','3KUqX8O4gabYhAorWq5toe':'Filip Piotr Makara','0CQY6JUvgOWr03z8v3Gd1y':'Filip Piotr Makara'}
filips_playlists_4 = {'5SjP9Ce1MjEOlNCiyToT2w':'Filip Piotr Makara','0IdKrjixDO3YuNK3FQaSP8':'Filip Piotr Makara','3Tq5Wd7oodEjwx0XnM73sd':'Filip Piotr Makara','0b3cP9e22uqakHwLSiHX7U':'Filip Piotr Makara'}
santiago_playlists_1 = {'3fkfTPWcPvTIz3rErPIJXP':'Santiago M. Gonzalez','41Mu7eCpeMhbzEFJlHE79O':'Santiago M. Gonzalez','51M0Iao8jzMG01Cc36DVcX':'Santiago M. Gonzalez','6fyIQ1KW6EHP3Yq47DFiIA':'Santiago M. Gonzalez','47LL6xPsiqgWu67XISl1x0':'Santiago M. Gonzalez'}
marta_1 = {'3lPnQ6MjFIwJk1idZC6E3y':'magweda1','6p8N0nuVkYp3n2rd75pjw8':'magweda1','3bjckb1HhGadSz7lkPwsAd':'magweda1','4mB2ChACmwXo1Z3oRXGaS7':'magweda1'}
makii_1 = {'37i9dQZF1DXaS0Bb5SeHIi':'Spotify','1eScuEm7ykdEfCafxjJPSE':'Scherz Ad','37i9dQZF1DX4AyFl3yqHeK':'Spotify'}
makii_2 = {'37i9dQZF1DX0XUsuxWHRQd':'Spotify'}
makii_3 = {'1Ep7ZNKybvRCXNsq9Jm6M3':'cyborgtemplar'}
makii_4 = {'06c5b4umxE7T4jJhGWPiBx':'Scout Wölfli','7usWuho4J9yEEk5VucM8rj':'Makii Oo','7rZgiJ8fpaFdocV23i1vVD':'Makii Oo','37i9dQZF1DWVzZlRWgqAGH':'Spotify'}
makii_5 ={'2bnxJ4zcQ42AC9cAZoqokl':'Makii Oo','66U2FUA2A4C9ErAAlrJtVs':'hotlemonwedge','37i9dQZF1DX7k3T9O9bscd':'Spotify','37i9dQZF1DX91UQmVbQYyN':'Spotify','5aesYspSWEh5zbu4BgVkdI':'Makii Oo','37i9dQZF1DX30w0JtSIv4j':'Spotify'}
alex_1 = {'6d3gdmWMqYi98lju6ysGNJ':'hotlemonwedge','72Bejw30fD0LovZyZ8LZIB':'hotlemonwedge','5XMGl3Kr1kRrVGfAqUMbF6':'hotlemonwedge','1nZN73yFATOS0zCYpjqE5H':'hotlemonwedge','4CubQTXmYk4Sq1zmNyV4Qv':'hotlemonwedge'}
alex_2 = {'7r82cJ0EwkQAWCr3tYQlAU':'hotlemonwedge'}
alex_3 = {'3xDez7ilwFjZfOMIBzK8MT':'hotlemonwedge','5mLJZ0RXujXRjhnE7jnIL2':'hotlemonwedge','1jRZY7xVO6u9IKvvBRNlyS':'hotlemonwedge'}
alex_4 = {'2KSTDM6nRl0IsYBPFA4mLz':'hotlemonwedge','6wcFGsuqYQwczGRDvd5VhE':'samtompkinsuk','1SCM5FgdvJ9Fof1QX4n9NY':'hotlemonwedge'}
various_pride = {'37i9dQZF1DX59HcpGmPXYR':'Spotify','37i9dQZF1DX5TMFhaZc9ov':'Spotify'}
various_country = {'37i9dQZF1DX1lVhptIYRda':'Spotify','37i9dQZF1DWTkxQvqMy4WW':'Spotify','37i9dQZF1DWZBCPUIUs2iR':'Spotify','37i9dQZF1DX7aUUBCKwo4Y':'Spotify','37i9dQZF1DWYnwbYQ5HnZU':'Spotify'}
various_reggae = {'37i9dQZF1DXbwoaqxaoAVr':'Spotify','37i9dQZF1DXbSbnqxMTGx9':'Spotify','37i9dQZF1DWWDLvD3P5wOG':'Spotify','37i9dQZF1DWYtKpmml7moA':'Spotify','37i9dQZF1DXa8n42306eJB':'Spotify'}
various_jazz = {'37i9dQZF1DX7YCknf2jT6s':'Spotify','37i9dQZF1DWTR4ZOXTfd9K':'Spotify','37i9dQZF1DWVmPb8C2FmNP':'Spotify','37i9dQZF1DXdwTUxmGKrdN':'Spotify','37i9dQZF1DX76YsWjvbz9I':'Spotify'}
various_classical = {'37i9dQZF1DWV0gynK7G6pD':'Spotify','37i9dQZF1DWXtBjoO4Thyy':'Spotify','37i9dQZF1DWXQD9LUA72yg':'Spotify','37i9dQZF1DWT9NPKiCMQns':'Spotify','37i9dQZF1DWZf52HmhYw49':'Spotify'}
various_funk = {'37i9dQZF1DWUS3jbm4YExP':'Spotify','37i9dQZF1DXcAiMO5AYxE8':'Spotify','37i9dQZF1DX8f5qTGj8FYl':'Spotify','37i9dQZF1DX70TzPK5buVf':'Spotify','37i9dQZF1DX2zVOKolQSzi':'Spotify'}
various_folk = {'37i9dQZF1DWYV7OOaGhoH0':'Spotify','37i9dQZF1DX9crXQ0wuuXE':'Spotify','37i9dQZF1DX504r1DvyvxG':'Spotify','37i9dQZF1DWVl5gPCRkquk':'Spotify','37i9dQZF1DWXJyjYpHunCf':'Spotify'}
various_latino = {'37i9dQZF1DX10zKzsJ2jva':'Spotify','37i9dQZF1DXbLMw3ry7d7k':'Spotify','37i9dQZF1DX65py6HnnlE1':'Spotify','37i9dQZF1DWZoF06RIo9el':'Spotify','37i9dQZF1DWSDautEwAKZB':'Spotify'}
various_ambient = {'37i9dQZF1DX9c7yCloFHHL':'Spotify','37i9dQZF1DWUrPBdYfoJvz':'Spotify','37i9dQZF1DXdgq0lsW8B2O':'Spotify','37i9dQZF1DWZ7mSWCFIT7v':'Spotify','37i9dQZF1DX8C585qnMYHP':'Spotify'}
various_blues = {'37i9dQZF1DX8QB9Ys2nV17':'Spotify'}
various_blues1 = {'37i9dQZF1DWZprlljz35Gr':'Spotify','37i9dQZF1DX7Y7BqFok9IQ':'Spotify'}
various_blues2 = {'37i9dQZF1DXcnkReojaCnV':'Spotify'}
various_soul = {'37i9dQZF1DX4SBhb3fqCJd':'Spotify','37i9dQZF1DX6VDO8a6cQME':'Spotify','37i9dQZF1DWTx0xog3gN3q':'Spotify','37i9dQZF1DWULEW2RfoSCi':'Spotify','37i9dQZF1DWT7oUl2XAhgF':'Spotify'}
various_metal = {'37i9dQZF1DWTcqUzwhNmKv':'Spotify','37i9dQZF1DX5J7FIl4q56G':'Spotify','37i9dQZF1DWWOaP4H0w5b0':'Spotify','37i9dQZF1DWXNFSTtym834':'Spotify','37i9dQZF1DXakaomPRkkDa':'Spotify'}
various_punk = {'37i9dQZF1DXd6tJtr4qeot':'Spotify','37i9dQZF1DXa9wYJr1oMFq':'Spotify','37i9dQZF1DX9wa6XirBPv8':'Spotify','37i9dQZF1DX3LDIBRoaCDQ':'Spotify','37i9dQZF1DX7WJ4yDmRK8R':'Spotify'}
gaussian_curve_inspired = {'37i9dQZF1DWZdlSSSctCmk':'Spotify','37i9dQZF1E4lAQFxLBkvBh':'Spotify'}
decades = {'37i9dQZF1DWXeOurMWUII5':'Spotify','37i9dQZF1DWTWEW1zqSeEj':'Spotify','37i9dQZF1DXdCc7Q1hwtuv':'Spotify','37i9dQZF1DX802IXCAaWtY':'Spotify','37i9dQZF1DWWUPe5aPjWhG':'Spotify','37i9dQZF1DWYfQ0uxBYM90':'Spotify'}
random_dictionary = {"6HEegfWHhUcytwQFAm1QbK" : "Frank Ocean" }
techno_dictionary = {"4RM93CShh3QYRKPnHmChSm" : "Krzysztof Kwiatkowski", "1jslqah910zhDryfFDLq7j" : "Krzysztof Kwiatkowski" }
hiphop_dictionary = {"37i9dQZF1DX2XmsXL2WBQd" : "Spotify"}
punk_dictionary = {"37i9dQZF1DXd6tJtr4qeot" : "Spotify"}
punk2 = {"37i9dQZF1EQqlvxWrOgFZm" : "Spotify" , "37i9dQZF1DX3MU5XUozve7": "Spotify"}
pop_hits = {'37i9dQZF1DWX7rdRjOECPW':'Spotify','37i9dQZF1DX3WvGXE8FqYX':'Spotify','37i9dQZF1DWWvoJqVv7uOD':'Spotify','37i9dQZF1DX7C2BlYJgCS5':'Spotify','37i9dQZF1DX2cNqJ4LgCMf':'Spotify','37i9dQZF1DX0s5kDXi1oC5':'Spotify','37i9dQZF1DXbYM3nMM0oPk':'Spotify'}
top_hits = {'37i9dQZEVXbMDoHDwVN2tF':'Spotify','37i9dQZF1ELYYzWaAD67i5':'Spotify','37i9dQZF1Etapvneoc5oia':'Spotify','37i9dQZF1DWUZv12GM5cFk':'Spotify','37i9dQZF1DXe2bobNYDtW8':'Spotify','37i9dQZF1Ejg7HCBbS3gRp':'Spotify'}
velvey_dict_5 = {'6l1RZerhJ0PF2Cnc9kivMp':'velvey','18V3QFlzjSbqyqAloYo3kh':'velvey'}
kk_1 = {'0KJUi7mAOTsUePrdX6r0ox':'Krzysztof Kwiatkowski','14qXDJrwuulLf3exANLFoO':'Krzysztof Kwiatkowski','0aqp4lz9ycEYQvN10gAUIa':'Krzysztof Kwiatkowski','1p280xCnCPO4PFpVfcgq98':'Krzysztof Kwiatkowski','33NMaOSGPBjd3LIBLr4pAE':'Krzysztof Kwiatkowski'}
kk_2 = {'23PD93UC3JGg9o9U6UyeJz':'Krzysztof Kwiatkowski','1UVmaFVxuikDCQV6SdlZFG':'Krzysztof Kwiatkowski','3LXBcw9KcSlLT1o2AtDgFg':'Krzysztof Kwiatkowski','4Smw6f7rJzvP6mKn1HJ2rG':'Krzysztof Kwiatkowski','63VD8EO5G5fVWJNh6FMCyW':'Krzysztof Kwiatkowski'}

## Add song titles and artist names to the dataframe

We have our dataframe prepared and saved in the Data folder, so we'll just show how it works here - the song names and artists will be necessary later for the song recommender, that's why we needed to scrape them from Spotify and later add to our dataframe.
<br><br>
The operation of getting the song titles and artist names from the Spotify API is very expensive timewise and it might get your Spotify account suspended, so we recommend using our prepared song database from the csv file

#read dataframe with songs audiofeatures 
#the dataframe is called cleaned because all duplicates have been removed, after creating it

songs = pd.read_csv("Data/songs_audio_features_cleaned.csv")
songs

#create separate dataframe with the songs ids and save to csv

song_ids = pd.DataFrame({'id': songs['id']})
song_ids.to_csv('Data/song_ids.csv', index=False)

#getting song titles for corresponding song IDs

song_titles = []

for song_id in list(song_ids['id']):
    song_titles.append(sp.track(song_id, market='GB')['name'])

#getting atist names for corresponding song IDs

artist_names = []

for song_id in list(song_ids['id']):
    artist_names.append(sp.track(song_id, market='GB')['album']['artists'][0]['name'])

#adding songs and artists to the ids

song_ids.insert(loc=0,column="title",value= pd.Series(song_titles))
song_ids.insert(loc=1,column="artist",value= pd.Series(artist_names))

song_ids_titles_and_artists = pd.read_csv('Data/song_ids_titles_and_artist_names.csv')

#drop unnecessary columns - the columns which will not be relevant for the clustering later
#concatenate the audio features with ids, songs and artist names

songs.drop(columns=['type','track_href','analysis_url','duration_ms','time_signature','uri','id'], inplace=True)

final_df = pd.concat(objs=[song_ids_titles_and_artists,songs], axis=1)

## Find audio features for user input

This function will be used later in the recommender.

def find_features_input():
    user_input = input("Please enter a song name here: ")
    x = sp.search(q="user_song",limit=1,market="GB")['tracks']['items'][0]['external_urls']['spotify'][31:]
    y = sp.audio_features(x)
    #make the format of the data frame the sme as our song list (+ standardize with standard scaler)
    return pd.DataFrame(y)
